stock_scanner_sovereign/requirements.txt
fyers-apiv3>=3.0.0
certifi
selenium
webdriver-manager
pyotp
pandas
pandas-ta
fastapi
uvicorn
streamlit
python-dotenv
requests
yfinance==0.2.18
pyarrow
python-dateutil
nselib
plotly
nbformat
xlsxwriter
psycopg2-binary
sqlalchemy
polars
polars
